+++ changelog.md
@@ -139,3 +139,4 @@
 
 * Fixed guards aggression towards mobs
 * Fixed profession change not always switching clothes
+* Added Family Tree item to search